handle delete

+++ src/Dashboard/ProductTable.tsx
@@ -6,6 +6,7 @@ import ProductModel from "./ProductModel/ProductModel";
 
 import { useAppDispatch } from "@/Redux/hooks";
 import { deleteProduct } from "@/Redux/features/products/productsSlice";
+import Swal from "sweetalert2";
 
 const ProductTable = ({ product }: { product: TProductProps }) => {
   const [update, setUpdate] = useState(String);
@@ -16,7 +17,27 @@ const ProductTable = ({ product }: { product: TProductProps }) => {
   };
 
   const handleDelete = () => {
-    dispatch(deleteProduct(product._id));
+    Swal.fire({
+      title: "Are you sure?",
+      text: "do you want to delete it?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteProduct(product._id));
+        Swal.fire({
+          position: "top-end",
+          icon: "success",
+          title: "Your Product has been Deleted",
+          showConfirmButton: false,
+          timer: 1500,
+        });
+      }
+    });
+
     console.log("clicked");
   };
   return (
